+++ src/xGadgetron/cGadgetron/include/sirf/Gadgetron/gadget_lib.h
@@ -180,10 +180,14 @@ namespace sirf {
 		}
 	};
 
+	class ImageMessageWriter : public aGadget {
+
+	};
 	/**
 	\brief Class for MRIImageWriter gadget xml-definition generator.
 	*/
-	class IsmrmrdImgMsgWriter : public aGadget {
+	//class IsmrmrdImgMsgWriter : public aGadget {
+	class IsmrmrdImgMsgWriter : public ImageMessageWriter {
 	public:
 		static const char* class_name()
 		{
@@ -199,6 +203,30 @@ namespace sirf {
 			xml_script += " <slot>1022</slot>\n";
 			xml_script += " <dll>gadgetron_mricore</dll>\n";
 			xml_script += " <classname>MRIImageWriter</classname>\n";
+			xml_script += "</writer>\n";
+			return xml_script;
+		}
+	};
+
+	/**
+	\brief Class for DicomImageWriter gadget xml-definition generator.
+	*/
+	class DicomImageMessageWriter : public ImageMessageWriter {
+	public:
+		static const char* class_name()
+		{
+			return "DicomImageWriter";
+		}
+		virtual void set_property(const char* prop, const char* value) {}
+		virtual std::string value_of(const char* prop)
+		{
+			return std::string("");
+		}
+		virtual std::string vxml(const std::string& label) const {
+			std::string xml_script("<writer>\n");
+			xml_script += " <slot>1018</slot>\n";
+			xml_script += " <dll>gadgetron_dicom</dll>\n";
+			xml_script += " <classname>DicomImageWriter</classname>\n";
 			xml_script += "</writer>\n";
 			return xml_script;
 		}
@@ -604,6 +632,17 @@ namespace sirf {
 		static const char* class_name()
 		{
 			return "ImageFinishGadget";
+		}
+	};
+
+	class DicomFinishGadget : public Gadget {
+	public:
+		DicomFinishGadget() :
+			Gadget("DicomFinish", "gadgetron_dicom", "DicomFinishGadget")
+		{}
+		static const char* class_name()
+		{
+			return "DicomFinishGadget";
 		}
 	};