+++ src/xSTIR/cSTIR/include/sirf/STIR/stir_x.h
@@ -38,6 +38,7 @@ limitations under the License.
 
 #define MIN_BIN_EFFICIENCY 1.0e-20f
 //#define MIN_BIN_EFFICIENCY 1.0e-6f
+#define DYNAMIC_CAST(T, X, Y) T& X = dynamic_cast<T&>(Y)
 
 namespace sirf {
 
@@ -541,17 +542,27 @@ The actual algorithm is described in
 		public stir::IterativeReconstruction < Image3DF > {
 	public:
 		bool post_process() {
+			//std::cout << "in xSTIR_IterativeReconstruction3DF.post_process...\n";
 			if (this->output_filename_prefix.length() < 1)
 				this->set_output_filename_prefix("reconstructed_image");
 			return post_processing();
 		}
 		stir::Succeeded setup(sptrImage3DF const& image) {
+			//std::cout << "in xSTIR_IterativeReconstruction3DF.setup...\n";
 			return set_up(image);
 		}
 		void update(Image3DF &image) {
 			update_estimate(image);
 			end_of_iteration_processing(image);
 			subiteration_num++;
+		}
+		void update(STIRImageData& id)
+		{
+			update(id.data());
+		}
+		void update(stir::shared_ptr<STIRImageData> sptr_id)
+		{
+			update(*sptr_id);
 		}
 		int& subiteration() {
 			return subiteration_num;
@@ -563,32 +574,6 @@ The actual algorithm is described in
 			initial_data_filename = filename;
 		}
 	};
-
-	class xSTIR_OSMAPOSLReconstruction3DF :
-		public stir::OSMAPOSLReconstruction < Image3DF > {
-	public:
-		stir::Succeeded set_up(stir::shared_ptr<STIRImageData> sptr_id)
-		{
-			stir::Succeeded s = stir::Succeeded::no;
-			xSTIR_IterativeReconstruction3DF* ptr_r =
-				(xSTIR_IterativeReconstruction3DF*)this;
-			if (!ptr_r->post_process()) {
-				s = ptr_r->setup(sptr_id->data_sptr());
-				ptr_r->subiteration() = ptr_r->get_start_subiteration_num();
-			}
-			return s;
-		}
-		void update(STIRImageData& id)
-		{
-			((xSTIR_IterativeReconstruction3DF*)this)->update(id.data());
-		}
-		void update(stir::shared_ptr<STIRImageData> sptr_id)
-		{
-			update(*sptr_id);
-		}
-	};
-
-	typedef xSTIR_OSMAPOSLReconstruction3DF OSMAPOSLReconstruction3DF;
 
 	class xSTIR_OSSPSReconstruction3DF : public stir::OSSPSReconstruction < Image3DF > {
 	public:
@@ -652,6 +637,48 @@ The actual algorithm is described in
 		stir::shared_ptr<STIRImageData> _sptr_image_data;
 	};
 
+	class xSTIR_SeparableGaussianImageFilter : 
+		public stir::SeparableGaussianImageFilter<float> {
+	public:
+		//stir::Succeeded set_up(const STIRImageData& id)
+		//{
+		//	return virtual_set_up(id.data());
+		//}
+		//void apply(STIRImageData& id)
+		//{
+		//	virtual_apply(id.data());
+		//}
+		void set_fwhms_xyz(char xyz, float f)
+		{
+			stir::BasicCoordinate<3, float> v = get_fwhms();
+			switch (xyz) {
+			case 'z':
+				v[1] = f;
+				break;
+			case 'y':
+				v[2] = f;
+				break;
+			case 'x':
+				v[3] = f;
+			}
+			set_fwhms(v);
+		}
+		void set_max_kernel_sizes_xyz(char xyz, int s)
+		{
+			stir::BasicCoordinate<3, int> v = get_max_kernel_sizes();
+			switch (xyz) {
+			case 'z':
+				v[1] = s;
+				break;
+			case 'y':
+				v[2] = s;
+				break;
+			case 'x':
+				v[3] = s;
+			}
+			set_max_kernel_sizes(v);
+		}
+	};
 }
 
 #endif